# docs/swagger.yaml
definitions:
...
  dto.HighStatePageResponse:
    properties:
      paging:
        $ref: '#/definitions/dto.PageResponse'
      results:
        description: Array of salt highstate results
        items:
          $ref: '#/definitions/model.HighState'
        type: array
    type: object
  dto.JIDPageResponse:
    properties:
      paging:
        $ref: '#/definitions/dto.PageResponse'
      results:
        description: Array of salt jid results
        items:
          $ref: '#/definitions/model.JID'
        type: array
    type: object
  dto.PageResponse:
    properties:
      count:
        description: Total number of items that match the query
        type: integer
      next:
        description: URL to the next page of results
        type: string
      num_pages:
        description: Total number of pages that match the query
        type: integer
      per_page:
        description: Total number of items per page that match the query
        type: integer
      previous:
        description: URL to the previous page of results
        type: string
    type: object
...
  github_com_jackc_pgtype.JSON:
    properties:
      bytes:
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/github_com_jackc_pgtype.Status'
    type: object
...
  github_com_jackc_pgtype.Status:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Undefined
    - "Null"
    - Present
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.HighState:
    properties:
      alterTime:
        type: string
      fullRet:
        $ref: '#/definitions/github_com_jackc_pgtype.JSON'
      fun:
        type: string
      id:
        type: string
      jid:
        type: string
      return:
        $ref: '#/definitions/github_com_jackc_pgtype.JSON'
      success:
        type: string
      user:
        type: string
    type: object
  model.JID:
    properties:
      alterTime:
        type: string
      jid:
        type: string
      load:
        $ref: '#/definitions/github_com_jackc_pgtype.JSON'
    type: object
...
info:
  contact:
    email: pmartin@gatech.edu
    name: API Support
    url: http://oit.gatech.edu
  description: This the Agartha API Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Agartha API
  version: "1.0"
paths:
...
  /api/v1/high_state:
    get:
      consumes:
      - application/json
      description: Get all HighStates
      parameters:
      - description: Page number of results to retrieve
        in: query
        name: page
        type: integer
      - description: restrict to X results
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HighStatePageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - HighState
  /api/v1/high_state/{id}:
    get:
      consumes:
      - application/json
      description: Get all HighStates for a specific id
      parameters:
      - description: id of the salt return item to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.HighState'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - HighState
  /api/v1/high_state/jid/{jid}:
    get:
      consumes:
      - application/json
      description: Get all HighStates for a specific jid
      parameters:
      - description: jid of the salt return item to retrieve
        in: path
        name: jid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.HighState'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - HighState
  /api/v1/jid:
    get:
      consumes:
      - application/json
      description: Get paginated jids with count and navigation links.
      parameters:
      - description: Filter items from this date (RFC3339 format).
        in: query
        name: since
        type: string
      - description: Filter items up to this date (RFC3339 format).
        in: query
        name: until
        type: string
      - description: Page number of results to retrieve
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JIDPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - JID
  /api/v1/jid/{jid}:
    get:
      consumes:
      - application/json
      description: Get one jid by JID
      parameters:
      - description: jid to return
        in: path
        name: jid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - JID
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"